API Cart.js added / Refactored all dependencies
