+++ server/api/products.js
@@ -44,39 +44,6 @@ router.get('/featured', async (req, res, next) => {
   }
 })
 
-router.put('/cart/add', (req, res, next) => {
-  try {
-    const addToCart = req.session.cart.push(req.body);
-    res.json(addToCart);
-  } catch (error) {
-    next(error);
-  }
-})
-
-router.put('/cart/delete', (req, res, next) => {
-  try {
-    let productToRemove = req.session.cart.find(cartItem => {
-      if (Number(cartItem.id) === Number(req.body.id)) {
-        return cartItem;
-      }
-    });
-    let indexOfProduct = req.session.cart.indexOf(productToRemove);
-    req.session.cart.splice(indexOfProduct, 1);
-    res.json(req.session.cart);
-  } catch (error) {
-    next(error)
-  }
-})
-
-router.get('/cart', (req, res, next) => {
-  try {
-    const cart = req.session.cart;
-    res.json(cart)
-  } catch (error) {
-    next(error)
-  }
-})
-
 // GetItemTypes
 router.get('/:itemType', async (req, res, next) => {
   try {
